+++ includes/functions/llms.functions.access.php
@@ -151,6 +151,61 @@ function llms_page_restricted( $post_id, $user_id = null ) {
 }
 
 /**
+ * Retrieve a message describing the reason why content is restricted
+ * Accepts an associative array of restriction data that can be retrieved from llms_page_restricted()
+ *
+ * This function doesn't handle all restriction types but it should in the future
+ * Currently it's being utilized for tooltips on lesson previews and some messages
+ * output during LLMS_Template_Loader handling redirects
+ *
+ * @param    array     $restriction  array of data from llms_page_restricted()
+ * @return   string
+ * @since    3.2.4
+ * @version  3.2.4
+ */
+function llms_get_restriction_message( $restriction ) {
+
+	$msg = __( 'You do not have permission to access to this content', 'lifterlms' );
+
+	switch ( $restriction['reason'] ) {
+
+		// this particular case is only utilized by lessons, courses do the check differently in the template
+		case 'course_time_period':
+			$course = new LLMS_Course( $restriction['restriction_id'] );
+			// if the start date hasn't passed yet
+			if ( ! $course->has_date_passed( 'start_date' ) ) {
+				$msg = $course->get( 'course_opens_message' );
+			} // course end date has passed
+			elseif ( $course->has_date_passed( 'end_date' ) ) {
+				$msg = $course->get( 'course_closed_message' );
+			}
+		break;
+
+		case 'enrollment_lesson':
+			$course = new LLMS_Course( $restriction['restriction_id'] );
+			$msg = $course->get( 'content_restricted_message' );
+		break;
+
+		case 'lesson_drip':
+			$lesson = new LLMS_Lesson( $restriction['restriction_id'] );
+			$msg = sprintf( _x( 'The lesson "%1$s" will be available on %2$s', 'lesson restricted by drip settings message', 'lifterlms' ), $lesson->get( 'title' ), $lesson->get_available_date() );
+		break;
+
+		case 'lesson_prerequisite':
+			$lesson = new LLMS_Lesson( $restriction['content_id'] );
+			$prereq_lesson = new LLMS_Lesson( $restriction['restriction_id'] );
+			$prereq_link = '<a href="' . get_permalink( $prereq_lesson->get( 'id' ) ) . '">' . $prereq_lesson->get( 'title' ) . '</a>';
+			$msg = sprintf( _x( 'The lesson "%1$s" cannot be accessed until the required prerequisite "%2$s" is completed.', 'lesson restricted by prerequisite message', 'lifterlms' ), $lesson->get( 'title' ), $prereq_link );
+		break;
+
+		default:
+
+	}
+
+	return apply_filters( 'llms_get_restriction_message', $msg, $restriction );
+}
+
+/**
  * Get a boolean out of llms_page_restricted for easy if checks
  * @param    int     $post_id   WordPress Post ID of the
  * @param    int     $user_id   optional user id (will use get_current_user_id() if none supplied)
@@ -317,7 +372,7 @@ function llms_is_post_restricted_by_time_period( $post_id, $user_id = null ) {
  * @return   bool|int          WP_Post ID of the membership if a restriction is found
  *                             false if no restrictions found
  * @since    3.0.0
- * @version  3.0.0
+ * @version  3.2.4 -- fix to handle post restricted to multiple memberships
  */
 function llms_is_post_restricted_by_membership( $post_id, $user_id = null ) {
 
@@ -341,7 +396,32 @@ function llms_is_post_restricted_by_membership( $post_id, $user_id = null ) {
 
 	if ( 'yes' === $restricted && $memberships && is_array( $memberships ) ) {
 
-		return absint( array_shift( $memberships ) );
+		// if no user, return the first membership from the array as the restriction id
+		if ( ! $user_id ) {
+
+			$restriction_id = array_shift( $memberships );
+
+		} // we have a user to check against each membership, user must be in at least one membership
+		else {
+
+			$student = new LLMS_Student( $user_id );
+
+			// loop through the memberships
+			foreach ( $memberships as $mid ) {
+
+				// set this as the restriciton id
+				$restriction_id = $mid;
+
+				// once we find the student has access break the loop
+				// this will be the restriciton that the template loader will check against later
+				if ( $student->has_access( $mid ) ) {
+					break;
+				}
+			}
+
+		}
+
+		return absint( $restriction_id );
 
 	}